Split test elements off into separate test mod
