+++ gradle.properties
@@ -1,3 +1,6 @@
+org.gradle.parallel=true
+fabric.loom.multiProjectOptimisation=true
+
 # `_DEV_` - Uncomment to compile against different version than what is required during runtime/production
 
 # Mod
