+++ testmod/common/build.gradle
@@ -1,0 +1,5 @@
+apply from: rootProject.file('scripts/common.gradle')
+
+dependencies {
+    compileOnly project(path: ':common', configuration: 'namedElements')
+}
