+++ testmod/common/src/main/java/xyz/apex/minecraft/testmod/common/block/TestFluidLoggingBlock.java
@@ -1,0 +1,42 @@
+package xyz.apex.minecraft.testmod.common.block;
+
+import net.minecraft.core.BlockPos;
+import net.minecraft.core.Direction;
+import net.minecraft.world.level.BlockGetter;
+import net.minecraft.world.level.block.Blocks;
+import net.minecraft.world.level.block.StairBlock;
+import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.level.block.state.properties.BlockStateProperties;
+import net.minecraft.world.level.block.state.properties.Half;
+import net.minecraft.world.level.block.state.properties.StairsShape;
+import net.minecraft.world.phys.shapes.CollisionContext;
+import net.minecraft.world.phys.shapes.VoxelShape;
+import xyz.apex.minecraft.apexcore.common.lib.component.block.BaseBlockComponentHolder;
+import xyz.apex.minecraft.apexcore.common.lib.component.block.BlockComponentRegistrar;
+import xyz.apex.minecraft.apexcore.common.lib.component.block.types.BlockComponentTypes;
+
+public final class TestFluidLoggingBlock extends BaseBlockComponentHolder
+{
+    public TestFluidLoggingBlock(Properties properties)
+    {
+        super(properties);
+    }
+
+    @Override
+    protected void registerComponents(BlockComponentRegistrar registrar)
+    {
+        registrar.register(BlockComponentTypes.WATERLOGGED);
+    }
+
+    @Override
+    public VoxelShape getShape(BlockState blockState, BlockGetter level, BlockPos pos, CollisionContext context)
+    {
+        return Blocks.OAK_STAIRS
+                .defaultBlockState()
+                .setValue(StairBlock.FACING, Direction.EAST)
+                .setValue(StairBlock.HALF, Half.BOTTOM)
+                .setValue(StairBlock.WATERLOGGED, blockState.getValue(BlockStateProperties.WATERLOGGED))
+                .setValue(StairBlock.SHAPE, StairsShape.STRAIGHT)
+                .getShape(level, pos, context);
+    }
+}
