+++ testmod/common/src/main/java/xyz/apex/minecraft/testmod/common/block/TestHorizontalBlock.java
@@ -1,0 +1,19 @@
+package xyz.apex.minecraft.testmod.common.block;
+
+import xyz.apex.minecraft.apexcore.common.lib.component.block.BaseBlockComponentHolder;
+import xyz.apex.minecraft.apexcore.common.lib.component.block.BlockComponentRegistrar;
+import xyz.apex.minecraft.apexcore.common.lib.component.block.types.BlockComponentTypes;
+
+public final class TestHorizontalBlock extends BaseBlockComponentHolder
+{
+    public TestHorizontalBlock(Properties properties)
+    {
+        super(properties);
+    }
+
+    @Override
+    protected void registerComponents(BlockComponentRegistrar registrar)
+    {
+        registrar.register(BlockComponentTypes.HORIZONTAL_FACING);
+    }
+}
