+++ testmod/fabric/build.gradle
@@ -1,0 +1,43 @@
+apply from: rootProject.file('scripts/fabric.gradle')
+
+loom {
+    runs {
+        data {
+            server()
+            ideConfigGenerated true
+            vmArg '-Dfabric-api.datagen'
+            vmArg "-Dfabric-api.datagen.output-dir=${project(':testmod-common').file('src/generated/resources')}"
+            vmArg "-Dfabric-api.datagen.modid=${MOD_ID}_testmod"
+        }
+    }
+
+    mods {
+        "${MOD_ID}" {
+            sourceSet project(':common').sourceSets.main
+            sourceSet sourceSets.main
+        }
+
+        "${MOD_ID}_testmod" {
+            sourceSet project(':testmod-common').sourceSets.main
+            sourceSet project(':testmod-fabric').sourceSets.main
+        }
+    }
+}
+
+dependencies {
+    implementation project(path: ':fabric', configuration: 'namedElements')
+    compileOnly project(path: ':testmod-common', configuration: 'namedElements')
+    include project(':fabric')
+}
+
+processResources {
+    from project(':testmod-common').sourceSets.main.resources
+}
+
+compileJava {
+    source project(':testmod-common').sourceSets.main.allSource
+}
+
+sourcesJar {
+    from project(':testmod-common').sourceSets.main.allJava
+}
