+++ testmod/fabric/src/main/java/xyz/apex/minecraft/testmod/fabric/TestModClientImpl.java
@@ -1,0 +1,9 @@
+package xyz.apex.minecraft.testmod.fabric;
+
+import org.jetbrains.annotations.ApiStatus;
+import xyz.apex.minecraft.testmod.common.TestModClient;
+
+@ApiStatus.Internal
+public final class TestModClientImpl implements TestModClient
+{
+}
