+++ testmod/fabric/src/main/java/xyz/apex/minecraft/testmod/fabric/entrypoint/TestModDataGeneratorEntrypoint.java
@@ -1,0 +1,17 @@
+package xyz.apex.minecraft.testmod.fabric.entrypoint;
+
+import net.fabricmc.fabric.api.datagen.v1.DataGeneratorEntrypoint;
+import net.fabricmc.fabric.api.datagen.v1.FabricDataGenerator;
+import org.jetbrains.annotations.ApiStatus;
+import xyz.apex.minecraft.apexcore.common.lib.resgen.ApexDataProvider;
+import xyz.apex.minecraft.testmod.common.TestMod;
+
+@ApiStatus.Internal
+public final class TestModDataGeneratorEntrypoint implements DataGeneratorEntrypoint
+{
+    @Override
+    public void onInitializeDataGenerator(FabricDataGenerator generator)
+    {
+        ApexDataProvider.register(TestMod.ID, func -> generator.createPack().addProvider(func::apply));
+    }
+}
