+++ testmod/fabric/src/main/java/xyz/apex/minecraft/testmod/fabric/entrypoint/package-info.java
@@ -1,0 +1,9 @@
+@MethodsReturnNonnullByDefault
+@FieldsAreNonnullByDefault
+@ParametersAreNonnullByDefault
+package xyz.apex.minecraft.testmod.fabric.entrypoint;
+
+import net.minecraft.FieldsAreNonnullByDefault;
+import net.minecraft.MethodsReturnNonnullByDefault;
+
+import javax.annotation.ParametersAreNonnullByDefault;
