+++ testmod/fabric/src/main/resources/META-INF/services/xyz.apex.minecraft.testmod.common.TestModClient
@@ -1,0 +1,1 @@
+xyz.apex.minecraft.testmod.fabric.TestModClientImpl
