+++ testmod/fabric/src/main/resources/fabric.mod.json
@@ -1,0 +1,85 @@
+{
+    "schemaVersion": 1,
+    "id": "${MOD_ID}_testmod",
+    "version": "${MOD_VERSION}",
+    "name": "${MOD_NAME} - TestMod",
+    "description": "TestMod for ${MOD_NAME}",
+    "authors": [
+        {
+            "name": "ApexStudios",
+            "contact": {
+                "homepage": "https://apexstudios.dev",
+                "issues": "https://discord.apexstudios.dev",
+                "sources": "https://github.com/ApexStudios-Dev/"
+            }
+        },
+        {
+            "name": "Apex - [Founder & Lead Programmer]",
+            "contact": {
+                "homepage": "https://apexmodder.xyz",
+                "issues": "https://discord.apexstudios.dev",
+                "sources": "https://github.com/ApexStudios-Dev/"
+            }
+        }
+    ],
+    "icon": "logo.png",
+    "contact": {
+        "homepage": "https://apexstudios.dev/",
+        "issues": "https://discord.apexstudios.dev",
+        "sources": "https://github.com/ApexStudios-Dev/"
+    },
+    "contributors": [
+        {
+            "name": "FantasyGaming (Viscerus) - [Co-Founder & Artist]",
+            "contact": {
+                "homepage": "https://bio.link/Viscerus",
+                "issues": "https://discord.apexstudios.dev",
+                "sources": "https://github.com/Viscerus"
+            }
+        },
+        {
+            "name": "RudySPG - [Web-Developer, Supporter, Beta Tester]",
+            "contact": {
+                "homepage": "https://github.com/RudySPG",
+                "sources": "https://github.com/RudySPG"
+            }
+        },
+        {
+            "name": "TobiSPG - [Supporter & Beta Tester]",
+            "contact": {
+                "homepage": "https://lexzandersaplan.com/",
+                "sources": "https://github.com/zanderlx"
+            }
+        }
+    ],
+    "license": "${MOD_LICENSE}",
+    "environment": "*",
+    "mixins": [ ],
+    "entrypoints": {
+        "main": [
+            "xyz.apex.minecraft.testmod.fabric.entrypoint.TestModModInitializer"
+        ],
+        "client": [
+            "xyz.apex.minecraft.testmod.fabric.entrypoint.TestModClientModInitializer"
+        ],
+        "fabric-datagen": [
+            "xyz.apex.minecraft.testmod.fabric.entrypoint.TestModDataGeneratorEntrypoint"
+        ]
+    },
+    "depends": {
+        "minecraft": "${MINECRAFT_FABRIC_VERSION_RANGE}",
+        "java": ">=${JAVA_VERSION}",
+        "fabric": "*",
+        "${MOD_ID}": "${MOD_VERSION}"
+    },
+    "custom": {
+        "modmenu": {
+            "links": {
+                "modmenu.discord": "https://discord.apexstudios.dev/"
+            },
+            "badges": [ "deprecated" ],
+            "update_checker": false,
+            "parent": "${MOD_ID}"
+        }
+    }
+}
